PBDS

+++ content/data-structures/HashMap.h
@@ -9,13 +9,13 @@
  */
 #pragma once
 
-#include <bits/extc++.h> /** keep-include */
+#include <ext/pb_ds/assoc_container.hpp> /// keep-include
 // To use most bits rather than just the lowest ones:
 struct chash { // large odd number for C
 	const uint64_t C = ll(4e18 * acos(0)) | 71;
 	ll operator()(ll x) const { return __builtin_bswap64(x*C); }
 };
-__gnu_pbds::gp_hash_table<ll,int,chash> h({},{},{},{},{1<<16});
+__gnu_pbds::gp_hash_table<ll,ll,chash> h({},{},{},{},{1<<16});
 
 /** For CodeForces, or other places where hacking might be a problem:
